Make methods private to Hoist
