+++ desktop/cmp/toolbar/Toolbar.js
@@ -10,7 +10,8 @@ import {hbox, vbox, fragment, filler} from '@xh/hoist/cmp/layout';
 import {overflowList, popover} from '@xh/hoist/kit/blueprint';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {Icon} from '@xh/hoist/icon';
-import {throwIf, filterConsecutiveToolbarSeparators} from '@xh/hoist/utils/js';
+import {throwIf} from '@xh/hoist/utils/js';
+import {filterConsecutiveToolbarSeparators} from '@xh/hoist/utils/impl';
 import {castArray} from 'lodash';
 import classNames from 'classnames';
 import PT from 'prop-types';
